Now result page will be rendered to show appropiate content upon submission.
- User will not be able to get points upon submitting solutions
 multiple times to the same problem.
- If user has already submitted a correct solution, then resubmitting
wrong solution would not deduct any point.
- Instead of using static messages in the result page, now it is capable
of generating messages dynamically depending on the results of the
testcases and user action.
- Publish permission status will now be shown on the home page and user will
get publish link automatically when the user has permission to do so,
else it will be hidden & an info icon will be display next to permission
status, which will show the required critaria to get the publish link.
- Footer size has been modified a bit.
- Code is now more PEP8 compliant.